+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.17.0.0
 
+* Remove `eqBootstrapWitnessRaw` and `BootstrapWitnessRaw`
+* Rename `bwSig` to `bwSignature` for `BootstrapWitness`
 * Remove `witVKeyBytes` and `eqWitVKeyRaw`
 * Remove `EqRaw` instance for `WitVKey`
 * Add `BoootstrapWitnessRaw` type